Make sure we finish sending the HTML payload before we exit the process
I encountered a page which responded successfully but all of the hydrogen HTML was missing.
It just had the boilerplate around it.

What I guessing happened is that since `process.send` is async, with a sufficiently large
payload and race condition, `process.exit(0)` was being called before it finished sending.