Fix test failing because the homeserver backfilled the first instead of the last message
All of the messages were at the same time, so the homeserver, just returned the first one
but in reality, it's a series of messages that happen one after the other so it should return
the last one.

We adjust the timestamps so they actually appear as a series. Although we think
about changing the sort on the date tiebreak when looking backwards in Synapse.