Fix `DisabledComposerView` description being overriden by powerlevel changes
We added the description in https://github.com/matrix-org/matrix-public-archive/pull/54
but then made a fix to one of the after render errors, https://github.com/matrix-org/matrix-public-archive/pull/57,
which exposed that the description is then overriden when the powerlevel changes.

We now just override the getters directly instead of the internal fields
which are overwritten internally.