Remove libera.chat as a default since their rooms are not accessible in the archive
The history visibility in Libera rooms is set to `join` which means it's not
accessible in the archive at all. Instead of leading a bunch of people to
`403 forbidden`, we can just remove from the default list.

The default list was mostly just copied from the Element list of defaults.